Rename plugin folder from roof_classify into RoofClassify
