+++ .github/workflows/packager.yml
@@ -1,7 +1,7 @@
 name: "📦 Packager"
 
 env:
-  PROJECT_FOLDER: "roof_classify"
+  PROJECT_FOLDER: "RoofClassify"
   PYTHON_VERSION: 3.9
 
 on:
