+++ RoofClassify/gui/dlg_settings.ui
