+++ RoofClassify/gui/roof_classify_dialog_base.ui
