+++ RoofClassify/resources/i18n/plugin_translation.pro
