+++ RoofClassify/resources/i18n/roof_classify_fr.ts
