+++ RoofClassify/resources/images/icon.png
