+++ docs/development/environment.md
@@ -20,7 +20,7 @@ python -m pip install -U pip
 # install development tools
 python -m pip install -U -r requirements/development.txt
 # install external dependencies
-python -m pip install --no-deps -U -r requirements/embedded.txt -t roof_classify/embedded_external_libs
+python -m pip install --no-deps -U -r requirements/embedded.txt -t RoofClassify/embedded_external_libs
 # install pre-commit to respect common development guidelines
 pre-commit install
 ```
