+++ requirements/embedded.txt
@@ -3,7 +3,7 @@
 # Python packages required by the plugin and which are not included into the embedded Python in QGIS (mainly Windows).
 #
 # Typical command to install:
-# python -m pip install --no-deps -U -r requirements/embedded.txt -t roof_classify/embedded_external_libs
+# python -m pip install --no-deps -U -r requirements/embedded.txt -t RoofClassify/embedded_external_libs
 # -----------------------
 
 joblib==1.1.*
