+++ setup.cfg
@@ -3,9 +3,9 @@
 description-file = README.md
 
 [qgis-plugin-ci]
-plugin_path = roof_classify
+plugin_path = RoofClassify
 github_organization_slug = DINFO-UniFI
-project_slug = roof_classify
+project_slug = RoofClassify
 
 # -- Code quality ------------------------------------
 [flake8]
@@ -51,7 +51,7 @@ use_parentheses = True
 addopts =
     --junitxml=junit/test-results.xml
     --cov-config=setup.cfg
-    --cov=roof_classify
+    --cov=RoofClassify
     --cov-report=xml
     --cov-report=html
     --cov-append tests/
@@ -65,7 +65,7 @@ testpaths = tests
 [coverage:run]
 branch = True
 include =
-    roof_classify/*
+    RoofClassify/*
 omit =
     .venv/*
     *tests*
